pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gp-pdes-sparsecholesky"
version = "0.1.0"
description = "Sparse approximate-Cholesky factorization and GP-based PDE solvers (Python port of KoLesky.jl + PDEs-GP-KoleskySolver)"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.10"
authors = [
    { name = "Yifan Chen", email = "yifanc96@gmail.com" },
]
keywords = [
    "gaussian-process", "pde", "sparse-cholesky",
    "kernel-methods", "scientific-computing",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Mathematics",
]
dependencies = [
    "numpy>=1.24",
    "scipy>=1.10",
    "jax>=0.4.20",
    "threadpoolctl>=3.0",
]

[project.optional-dependencies]
gpu = [
    # JAX with CUDA — pick a version that matches your local CUDA toolkit.
    "jax[cuda12]>=0.4.20",
    # CuPy is optional; only used if you set KOLESKY_ENABLE_GPU_SPARSE=1.
    "cupy-cuda12x>=13.0",
]
dev = [
    "pytest>=7.0",
]

[project.urls]
Homepage = "https://github.com/yifanc96/GP-PDEs-SparseCholesky"
"Source" = "https://github.com/yifanc96/GP-PDEs-SparseCholesky"
"Paper (KoLesky)" = "https://arxiv.org/abs/2004.14455"
"Paper (PDE solver)" = "https://arxiv.org/abs/2304.01294"

[tool.setuptools.packages.find]
include = ["kolesky*"]
exclude = ["examples*", "tests*"]
